pyproject.toml
[project]
name = "bhw-dc-bot"
version = "0.1.0"
description = "Helferlein für die Bens Hardware Discord Community"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "discord-py>=2.6.4",
    "google-genai>=1.60.0",
    "levenshtein>=0.27.3",
    "pytesseract>=0.3.13",
]
